requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.2

# Computer Vision & OCR
opencv-python>=4.7.0
pytesseract>=0.3.10
ultralytics>=8.0.0  # For YOLO

# Vector Database
pinecone-client>=2.2.2

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
